Split ComponentSettingsDialogViewModel import helpers into a partial (fix 500-line hard limit)
#632 pushed ComponentSettingsDialogViewModel.cs to 507 lines, tripping the
NoProductionFileExceedsHardLimit architecture test (hard limit 500) — which was
red on main and blocking every PR. Move the S-parameter import helpers
(ReconcilePortNamesAsync, BuildImportStatus, FindImporter, LooksLikeLumericalTxt)
into a cohesive partial file. Pure code move, no behavior change; main file is
now 408 lines.
